Coldfix Sprint 2

See merge request ppl-fasilkom-ui/2020/ppl-c/diskominfo-depok---masif/abang-abang-masif!41

+++ user/views.py
@@ -4,70 +4,74 @@ import re
 from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render, redirect
 from django.utils.datastructures import MultiValueDictKeyError
+from django.contrib import messages
+from django.contrib.sites.shortcuts import get_current_site
 
-from account.models import UserProfile
+from account.models import UserProfile, Account
+from admin.mailing import send_verification_email
+from user.models import UserVerificationList
 from user.forms import EditUserProfileForm, CVForm, ProfilePictureForm
+from .user_registration_form import UserRegistrationForm
+from .token import generate_user_token
 
+URL_USER_DASHBOARD = '/user/dashboard/'
+ERROR_PAGE_NOT_FOUND = 'ERROR 404 Page not found'
 
 def account_is_user(request):
     try:
-        return request.user.is_authenticated and request.user.is_user and request.user.userprofile is not None
+        return (request.user.is_authenticated
+                and request.user.is_user
+                and request.user.userprofile is not None)
     except UserProfile.DoesNotExist:
         return False
 
 
 def upload_profile_picture(request):
-    if account_is_user(request):
-        if request.method == 'POST':
-            form = ProfilePictureForm(request.POST, request.FILES)
-            if form.is_valid():
-                request.user.profile_picture = request.FILES['profile_picture']
-                request.user.save()
-            return redirect('/user/dashboard/')
-    return HttpResponse('ERROR 404 Page not found')
+    if account_is_user(request) and request.method == 'POST':
+        form = ProfilePictureForm(request.POST, request.FILES)
+        if form.is_valid():
+            request.user.profile_picture = request.FILES['profile_picture']
+            request.user.save()
+        return redirect(URL_USER_DASHBOARD)
+    return HttpResponse(ERROR_PAGE_NOT_FOUND)
 
 
 def upload_cv(request):
-    if account_is_user(request):
-
-        if request.method == 'POST':
-            form = CVForm(request.POST, request.FILES)
-            # form.instance.user = request.user
-            if form.is_valid():
-                try:
-                    request.user.userprofile.cv = request.FILES['cv']
-                    request.user.userprofile.save()
-                except MultiValueDictKeyError:
-                    return HttpResponse('ERROR no file sent')
-            return redirect('/user/dashboard/')
-    return HttpResponse('ERROR 404 Page not found')
+    if account_is_user(request) and request.method == 'POST':
+        form = CVForm(request.POST, request.FILES)
+        if form.is_valid():
+            try:
+                request.user.userprofile.cv = request.FILES['cv']
+                request.user.userprofile.save()
+            except MultiValueDictKeyError:
+                return HttpResponse('ERROR no file sent')
+        return redirect(URL_USER_DASHBOARD)
+    return HttpResponse(ERROR_PAGE_NOT_FOUND)
 
 
 def delete_cv(request):
-    if account_is_user(request):
-        if request.method == 'POST':
-            request.user.userprofile.cv.delete()
-    return HttpResponse('ERROR 404 Page not found')
+    if account_is_user(request) and request.method == 'POST':
+        request.user.userprofile.cv.delete()
+    return HttpResponse(ERROR_PAGE_NOT_FOUND)
 
 
 def user_dashboard(request):
     if account_is_user(request):
 
         return render(request, 'user/user-dashboard.html',
-                      {'form_pp': ProfilePictureForm(), 'user': request.user, 'form_cv': CVForm()})
-    else:
-        return redirect('/user/login/')
+                      {'form_pp': ProfilePictureForm(),
+                       'user': request.user, 'form_cv': CVForm()})
+    return redirect('/user/login/')
 
 
 def born_date_validator(post):
-    # date_format = '%Y-%m-%d'
     date = str(post['born_date'])
     date_format = '%d/%m/%Y'
     try:
         date_obj = datetime.datetime.strptime(date, date_format)
         print(date_obj)
         return {'result': True, 'message': 'success'}
-    except (ValueError, TypeError) as e:
+    except (ValueError, TypeError):
         return {'result': False, 'message': 'Tanggal lahir salah'}
 
 
@@ -113,7 +117,9 @@ def user_edit_profile(request):
                     request.user.phone = request.POST['phone']
                     profile.born_city = request.POST['born_city']
 
-                    date_obj = datetime.datetime.strptime(request.POST['born_date'], '%d/%m/%Y')
+                    date_obj = datetime.datetime.strptime(
+                        request.POST['born_date'],
+                        '%d/%m/%Y')
                     profile.born_date = date_obj
                     profile.address = request.POST['address']
                     profile.sex = request.POST['sex']
@@ -127,8 +133,11 @@ def user_edit_profile(request):
                 return HttpResponse(data_valid['message'])
             return HttpResponse(form.errors)
         else:
-            born_date_in_database = str(request.user.userprofile.born_date).split('-')
-            shown_born_date = '{}/{}/{}'.format(born_date_in_database[2], born_date_in_database[1],
+            born_date_in_database = str(
+                request.user.userprofile.born_date
+                ).split('-')
+            shown_born_date = '{}/{}/{}'.format(born_date_in_database[2],
+                                                born_date_in_database[1],
                                                 born_date_in_database[0])
 
             current_user_data = {
@@ -139,12 +148,11 @@ def user_edit_profile(request):
                 'address': request.user.userprofile.address,
                 'institution': request.user.userprofile.institution,
                 'major': request.user.userprofile.major,
-                'phone': request.user.phone,
-
             }
 
             return render(request, 'user/user-edit-profile.html',
-                          {'form': EditUserProfileForm(current_user_data), 'request': request})
+                          {'form': EditUserProfileForm(current_user_data),
+                           'request': request})
     return redirect('/user/login/')
 
 
@@ -154,25 +162,20 @@ def list_of_lowongan_to_json_dict(data):
     }
 
     for lowongan in data:
-        result['data'].append(["Pending", '<a href="/cari-lowongan/detail-lowongan/">{}</a>'.format(lowongan.judul),
-                               lowongan.opd_foreign_key.name])
+        result['data'].append(
+            ["Pending",
+             '<a href="/cari-lowongan/detail-lowongan/">{}</a>'.format(
+                 lowongan.judul),
+             lowongan.opd_foreign_key.name])
     return result
 
 
 def get_all_lamaran_for_dashboard_table(request):
     if request.user.is_authenticated:
-        response = list_of_lowongan_to_json_dict(request.user.userprofile.lowongan_set.all())
+        response = list_of_lowongan_to_json_dict(
+            request.user.userprofile.lowongan_set.all())
         return JsonResponse(response)
-    else:
-        return HttpResponse('[ERROR] permission denied')
-
-
-from django.contrib import messages
-
-from account.models import Account, PelamarProfile
-from user.models import UserVerificationList
-from .user_registration_form import UserRegistrationForm
-from .token import generate_user_token
+    return HttpResponse('[ERROR] permission denied')
 
 def user_register(request):
     if request.method == 'POST':
@@ -191,10 +194,12 @@ def user_register(request):
                 password=password
             )
             new_account.save()
+            base_url = get_current_site(request).domain
+            verif_url = base_url+'/user/verification/'+secret
+            send_verification_email(verif_url, email)
             return render(
                 request,
-                'user/user_activation_link.html',
-                {'secret': secret}
+                'activation_link.html',
             )
     else:
         form = UserRegistrationForm()
@@ -208,7 +213,7 @@ def user_verification(request, token):
         email = user_from_verification_list.email
         phone = user_from_verification_list.phone
         password = user_from_verification_list.password
-    except:
+    except UserVerificationList.DoesNotExist:
         return redirect('/user/verification/404')
     new_user = Account.objects.create_user(email, password)
     new_user.name = user_name
